Support Kijiji ads using the new layout (fixes #6)
Some ads (but not all of them) use a different layout
which breaks the old scraper. The HTML of these new ads
contains a helpful JavaScript object with tons of ad
information, and it is used when possible. If this object
is not found, it is assumed that the ad is using the old
layout, and the page is scraped using the old method as
needed.

Additional changes:
* The listing date is converted to a JavaScript data object
  when it is scraped