Refactor, cleanup, and improve search and scrape functionality
* A new KijijiAd class has been created to provide an easy interface
to an ad's information and to facilitate scraping

* Kijiji's HTML search results page is now used instead of the RSS feed
to perform searches. This means that the returned information is
slightly different, but allows for more consistency (and paging, which
leads into the next point)

* The maximum and minimum number of desired search results can now be
specified. Results are no longer limited to 20 ads

* Ads retrieved from a search now follow the same structure as those
retrieved through scraping a URL. Information not able to be gleaned
from the search results page will simply be missing. KijijiAd.isScraped()
can be used to determine if all of its information has been retrieved,
and KijijiAd.scrape() can be used to retrieve it. Alternately,
scrapeResultDetails=true can be passed as a search option to perform this
operation automatically when retrieving search results (true is the
default).