Handle the case where it is not possible to suppress following redirects
This can happen if running the scraper from within a web browser.

This commit also introduces a minor optimization for pagination: the
search parameters are now only submitted once and the path to the first
page of results is saved. The stored path is then modified on requests for
subsequent pages.