Date handling (#58)
* Report errors returned by the mobile API

* Use ad 'start date' for API scraper (#57)

Kijiji differentiates between when an ad is created, modified, started,
and ending. The start date is what is used in search results and ad
listings. Ads can be "refreshed" to appear new, which simply moves the
start date forward.

The HTML scraper does not need to be changed since it only has access to
the date value in the markup, which is the start date.

* Add note to readme regarding delay between ad creation and visibility (#57)

* Bump version